requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.1
django-jazzmin==3.0.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.9.0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
